requirements.txt
gdown
torch
numpy
pafy
Pillow
einops
pytorch_lightning
youtube-dl==2020.12.2
cv2